Unblock textarea send during edit
